Fixed final merge conflicts and issues on the frontend
