+++ .gitignore
@@ -11,7 +11,7 @@ Full_DDL.sql
 
 # MERGING
 marketplace.jsx
-
+documentation.jsx
 
 # Distribution / packaging
 .Python
