+++ backend/uclapi/oauth/app_helpers.py
@@ -22,7 +22,7 @@ def generate_random_verification_code():
 def get_student_by_upi(upi):
     # Returns a StudentA or StudentB object by UPI
     if Lock.objects.all()[0].a:
-        students = StudentsA \
+        students = StudentsA
     else:
         students = StudentsB
