+++ frontend/src/components/documentation/Routes/Timetable/GetPersonalTimetable.jsx
@@ -5,7 +5,7 @@ import Table from './../../Table.jsx';
 import Cell from './../../Cell.jsx';
 
 // Code Generator 
-import * as RequestGenerator from 'Layout/Data/RequestGenerator.jsx';
+import * as RequestGenerator from 'Layout/data/RequestGenerator.jsx';
 
 let params = {
   "token": "uclapi-user-5d58c3c4e6bf9c-c2910ad3b6e054-7ef60f44f1c14f-a05147bfd17fdb",
